refactor(form): form ui component
affects: @sections/hero, @ui/form, @ui/modal
