--- frontend/fragments/form-hero/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@fragments/form-hero",
-  "version": "0.0.4",
-  "main": "src/index.tsx",
-  "license": "MIT",
-  "dependencies": {
-    "@atlantis-lab/layout": "^0.3.0",
-    "@ui/button": "^0.0.4",
-    "@ui/input": "^0.0.3",
-    "@ui/text": "^0.0.4",
-    "@ui/theme": "^0.1.0",
-    "react": "^16.13.1",
-    "react-intl": "^5.0.2"
-  }
-}
