--- frontend/fragments/header-modal-form/CHANGELOG.md
@@ -1,49 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## 0.0.4 (2020-10-05)
-
-### Bug Fixes
-
-- **hero:** change layout and box ([4bdf1ca](https://github.com/Atlantis-Academy/react-learn-landing/commit/4bdf1caea3bd0aecf863ff04d0926624ea644db2))
-- **hero:** change layouts, add docker ([86d4eff](https://github.com/Atlantis-Academy/react-learn-landing/commit/86d4eff0254c4e07cd31705cfc40a85b7e24b143))
-- **hero:** delete console.log, waves ([ce86fef](https://github.com/Atlantis-Academy/react-learn-landing/commit/ce86fef8afde6947f1e2d3ac9cfbce2072f6e9d7))
-- **input:** inputs working ([f830d30](https://github.com/Atlantis-Academy/react-learn-landing/commit/f830d30f11ff862a790d1b8a5d8829818da781ad))
-- **input:** theme for input, delete header-button ([a19a546](https://github.com/Atlantis-Academy/react-learn-landing/commit/a19a54617e06cd90a8afb66d00ce54100b6608d8))
-- **messages, name components:** fix all but not press Enter ([65131d8](https://github.com/Atlantis-Academy/react-learn-landing/commit/65131d8d5c6517db911f1eb4d5a347114d2183a5))
-- **modal:** rename onClose modal ([0e3436c](https://github.com/Atlantis-Academy/react-learn-landing/commit/0e3436c60a0091270ce878f688ba6e556ddc9be5))
-- **version:** update versions packages ([7110f03](https://github.com/Atlantis-Academy/react-learn-landing/commit/7110f034c05cef804ef32c0a4e943dba33b763c3))
-
-## 0.0.3 (2020-09-22)
-
-### Bug Fixes
-
-- **hero:** change layout and box ([4bdf1ca](https://github.com/Atlantis-Academy/react-learn-landing/commit/4bdf1caea3bd0aecf863ff04d0926624ea644db2))
-- **hero:** change layouts, add docker ([86d4eff](https://github.com/Atlantis-Academy/react-learn-landing/commit/86d4eff0254c4e07cd31705cfc40a85b7e24b143))
-- **hero:** delete console.log, waves ([ce86fef](https://github.com/Atlantis-Academy/react-learn-landing/commit/ce86fef8afde6947f1e2d3ac9cfbce2072f6e9d7))
-- **input:** inputs working ([f830d30](https://github.com/Atlantis-Academy/react-learn-landing/commit/f830d30f11ff862a790d1b8a5d8829818da781ad))
-- **input:** theme for input, delete header-button ([a19a546](https://github.com/Atlantis-Academy/react-learn-landing/commit/a19a54617e06cd90a8afb66d00ce54100b6608d8))
-- **messages, name components:** fix all but not press Enter ([65131d8](https://github.com/Atlantis-Academy/react-learn-landing/commit/65131d8d5c6517db911f1eb4d5a347114d2183a5))
-- **modal:** rename onClose modal ([0e3436c](https://github.com/Atlantis-Academy/react-learn-landing/commit/0e3436c60a0091270ce878f688ba6e556ddc9be5))
-
-## 0.0.2 (2020-08-24)
-
-### Bug Fixes
-
-- **deps:** change dependencies mask for modal ([c01a663](https://github.com/Atlantis-Academy/react-learn-landing/commit/c01a663e980e1aa4c87f47707d5c70ff2f257107))
-- **hero:** change dependencies button and input ([908b87f](https://github.com/Atlantis-Academy/react-learn-landing/commit/908b87fbac89a80c8012bc187c657400ee1fa723))
-- **hero:** change dependencies input ([b699e0d](https://github.com/Atlantis-Academy/react-learn-landing/commit/b699e0d7f087d6ed61e44043f6366daf9d96b341))
-- **hero:** changes width review ([ccf5d91](https://github.com/Atlantis-Academy/react-learn-landing/commit/ccf5d91910d007dd33460200a3b3aae21e720191))
-- **hero:** rename module close-cross and fragments ([6a5f803](https://github.com/Atlantis-Academy/react-learn-landing/commit/6a5f803041966ea03e00db1db21d049549c9a0ec))
-- **input:** add input component ([3147f00](https://github.com/Atlantis-Academy/react-learn-landing/commit/3147f004dd0e325d50254aa7594ca33852bf2f0f))
-- **version:** change version 0.0.0 ([5812a9c](https://github.com/Atlantis-Academy/react-learn-landing/commit/5812a9cc39d1c17a3de4e75eb7a06b68bd387789))
-- **version:** change version 0.0.1 ([dc59a12](https://github.com/Atlantis-Academy/react-learn-landing/commit/dc59a120ac423e2d0d557f282be696d93a832cf7))
-
-## 0.0.1 (2020-08-14)
-
-### Bug Fixes
-
-- **hero:** changes width review ([ccf5d91](https://github.com/Atlantis-Academy/react-learn-landing/commit/ccf5d91910d007dd33460200a3b3aae21e720191))
-- **version:** change version 0.0.0 ([5812a9c](https://github.com/Atlantis-Academy/react-learn-landing/commit/5812a9cc39d1c17a3de4e75eb7a06b68bd387789))
