--- frontend/fragments/header-modal-form/src/index.tsx
@@ -1,3 +1,0 @@
-import HeaderModalForm from './HeaderModalForm'
-
-export { HeaderModalForm }
