+++ frontend/sections/hero/package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "dependencies": {
     "@atlantis-lab/layout": "^0.3.0",
-    "@fragments/form-hero": "^0.0.4",
+    "@ui/form": "0.0.0",
     "@ui/background": "^0.0.2",
     "@ui/icons": "^0.1.0",
     "@ui/text": "^0.0.4",
