+++ frontend/sections/hero/src/Hero.tsx
@@ -2,7 +2,7 @@ import React                                           from 'react'
 import { Box, Column, Layout, Row }                    from '@atlantis-lab/layout'
 import { injectIntl }                                  from 'react-intl'
 
-import { Form }                                        from '@fragments/form-hero'
+import { Form }                                        from '@ui/form'
 import { Background, BackgroundBottom, BackgroundTop } from '@ui/background'
 import { Divider }                                     from '@ui/icons'
 import { Text }                                        from '@ui/text'
@@ -67,7 +67,7 @@ const Hero = ({ intl }) => {
                         {intl.formatMessage(messages.formTitle)}
                       </Text>
                       <Layout flexBasis={19} />
-                      <Form />
+                      <Form context='hero' textButton={intl.formatMessage(messages.button)} />
                       <Layout flexBasis={40} />
                     </Column>
                   </Box>
