+++ frontend/ui/form/src/Messages.tsx
@@ -3,10 +3,6 @@ import { defineMessages } from 'react-intl'
 import { name as scope }  from '../package.json'
 
 export default defineMessages({
-  button: {
-    id: `${scope}.button`,
-    defaultMessage: 'отправить заявку',
-  },
   name: {
     id: `${scope}.name`,
     defaultMessage: 'Имя',
@@ -15,16 +11,24 @@ export default defineMessages({
     id: `${scope}.email`,
     defaultMessage: 'Email',
   },
-  msg: {
-    id: `${scope}.msg`,
-    defaultMessage: 'Сообщение',
+  job: {
+    id: `${scope}.job`,
+    defaultMessage: 'Организация',
+  },
+  phone: {
+    id: `${scope}.phone`,
+    defaultMessage: 'Телефон',
   },
   success: {
     id: `${scope}.success`,
     defaultMessage: 'Спасибо! Мы скоро с вами свяжемся!',
   },
-  declined: {
-    id: `${scope}.declined`,
-    defaultMessage: 'Заполните все поля формы.',
+  required: {
+    id: `${scope}.required`,
+    defaultMessage: 'Необходимо заполнить все поля формы!',
+  },
+  message: {
+    id: `${scope}.message`,
+    defaultMessage: 'Сообщение',
   },
 })
