+++ frontend/ui/modal/package.json
@@ -6,7 +6,7 @@
   "dependencies": {
     "@atlantis-lab/layout": "^0.3.0",
     "@emotion/styled": "^10.0.27",
-    "@fragments/header-modal-form": "^0.0.4",
+    "@ui/form": "0.0.0",
     "@ui/button": "^0.0.4",
     "@ui/icons": "^0.1.0",
     "@ui/text": "^0.0.4",
